Add scores and functions fullGame and winner
Add playerScore and computerScore variables to keep score of the game.
Add relevant score each time within the gameRound function.
Make fullGame function that loops for five times to make five rounds of game, keeping score of each round.
Make winner function that calculates who is the winner at the end of the fifth round and declares it.